Add files via upload
This is a traffic detection and management prototype.
In a traffic point, this program with the help of cameras can detect and count the number of vehicles in each lane. 
Then instead of letting traffic in each lane pass for a pre-determined amount of time (90 seconds in general), each lane will be given different amount of time depending on the traffic density. In this way, everything is dynamically managed.